requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
decorator==4.4.2
docstring_parser==0.16
eval_type_backport==0.2.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
setuptools==78.1.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
together==1.5.18
torch==2.7.1
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vertexai==1.71.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
